Update DwellingManager to support polymorphic record display with different format modes
Updated the dwelling_manager.py file to work with our new class hierarchy. This will add the ability to display records in different formats using polymorphism.
Here's what we added to DwellingManager class:

A display mode property to track which format to use.
A method to switch between display modes.
A method to convert records to the appropriate display type.